--- README.md
@@ -1,10 +1,0 @@
-# Calculator
-Software Engineering Project: Assignment 1
-
-Group members:
-1. Adaora Arinze
-2. James Fenlon
-
-This is our main branch of the GitHub repository. Here is the most up to date and fully functioning calculator.
-Running the Calculator.java file will start our ptogram. You can then start by inputting your desired equation, or enter terminate to stop the session.
-Enjoy!!
